moved authtableinfo struct to dbauthenticators
